+++ src/entities/plugin.ts
@@ -3,6 +3,8 @@ import { Commit } from './commit';
 import { Context } from './state';
 import { ConfigurationOptions } from './configuration';
 
+export type PluginType = Plugin | CommitPlugin | StatePlugin;
+
 export abstract class Plugin {
     protected context: Context;
 
@@ -11,5 +13,12 @@ export abstract class Plugin {
     }
 
     public abstract async init(config: ConfigurationOptions): Promise<void>;
+}
+
+export abstract class CommitPlugin extends Plugin {
     public abstract async parse(commit: Commit, task: Task): Promise<void>;
+}
+
+export abstract class StatePlugin extends Plugin {
+    public abstract async modify(task: Task): Promise<void>;
 }