+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Important Changes
+Source code now under `MIT` license.
 ## Features
 ### Logging
 - Added base logging system a task & statuses [`ba1e7ee`](https://github.com/keindev/changelog-guru/commit/ba1e7eeb4ac0c12044cbbe4ba17337add9f4bef1)
@@ -13,6 +15,8 @@
 - **[API]**  Replace @octokit/rest by graphql [`aabcdbd`](https://github.com/keindev/changelog-guru/commit/aabcdbd20a00ceecf9c2bd89a0608e5c3e07b565) [`84aeb3b`](https://github.com/keindev/changelog-guru/commit/84aeb3bafa599e96c5ef6418d817289b716fbdc4)
 ### Subsections & Groups
 - **[Section]**  Added new SectionBlock type - Group [`e1e837c`](https://github.com/keindev/changelog-guru/commit/e1e837ccf41fd543af77bf2595bca97adfecb0d7)
+### Package info
+- **[Package]**  Add of receipt info about changes in package [`b2407bd`](https://github.com/keindev/changelog-guru/commit/b2407bdf737626f9800570591b21ccc3e67e6034)
 ### Others
 - **[Config]**  Use cosmiconfig for load config file [`833e950`](https://github.com/keindev/changelog-guru/commit/833e95021ad46296d52c1b70cbf21879377f9bc8)
 - **[Config]**  Rename config file [`a34b824`](https://github.com/keindev/changelog-guru/commit/a34b824b2b5cc152b586e5b6df7b374e3a27983c)